moved PreQuestionComponent to types
